api/requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
chex==0.0.8
clang==5.0
click==8.0.3
cycler==0.11.0
dm-tree==0.1.6
filelock==3.3.2
firebase-admin==5.0.3
Flask==2.0.2
flatbuffers==1.12
flax==0.3.6
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gunicorn==20.1.0
grpcio==1.41.1
h5py==3.1.0
huggingface-hub==0.0.19
idna==3.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jax==0.2.24
jaxlib==0.1.73
Jinja2==3.0.2
joblib==1.1.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
msgpack==1.0.2
numpy==1.19.5
oauthlib==3.1.1
opt-einsum==3.3.0
optax==0.0.9
packaging==21.2
Pillow==8.4.0
protobuf==3.19.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
PyYAML==6.0
pyzmq==19.0.2
regex==2021.10.23
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.46
scipy==1.7.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
tensorflow-hub==0.12.0
termcolor==1.1.0    
tokenizers==0.10.3
toolz==0.11.1
tqdm==4.62.3
transformers==4.12.2
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
